src/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
python-multipart==0.0.12
azure-identity==1.19.0
azure-storage-blob==12.24.0
azure-cosmos==4.7.0
azure-keyvault-secrets==4.9.0
azure-monitor-opentelemetry==1.6.5
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-httpx==0.49b2
PyJWT[crypto]==2.10.0
